Rename C++ directory `modmesh/view/` to `modmesh/pilot/` and the extension module name `_modmesh.view` to `_modmesh.pilot`
