+++ cpp/binary/pilot/pilot.cpp
