+++ cpp/modmesh/pilot/RAction.cpp
@@ -1,5 +1,3 @@
-#pragma once
-
 /*
  * Copyright (c) 2022, Yung-Yu Chen <yyc@solvcon.net>
  *
@@ -28,50 +26,22 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <modmesh/view/common_detail.hpp> // Must be the first include.
+#include <modmesh/pilot/RAction.hpp> // Must be the first include.
 
-#include <modmesh/modmesh.hpp>
-
-#include <Qt>
-#include <QWidget>
-#include <Qt3DWindow>
-
-#include <QByteArray>
-#include <QGeometryRenderer>
-
-#include <Qt3DCore/QBuffer>
-#include <Qt3DCore/QEntity>
-#include <Qt3DCore/QGeometry>
-#include <Qt3DCore/QAttribute>
-#include <Qt3DCore/QTransform>
-
-#include <Qt3DExtras/QDiffuseSpecularMaterial>
+#include <functional>
 
 namespace modmesh
 {
 
-class RStaticMesh
-    : public Qt3DCore::QEntity
+RAction::RAction(QString const & text, QString const & tipText, std::function<void(void)> callback, QObject * parent)
+    : QAction(text, parent)
 {
-
-public:
-
-    RStaticMesh(std::shared_ptr<StaticMesh> const & static_mesh, Qt3DCore::QNode * parent = nullptr);
-
-    void update_geometry(StaticMesh const & mh)
+    setStatusTip(tipText);
+    if (callback != nullptr)
     {
-        update_geometry_impl(mh, m_geometry);
+        connect(this, &QAction::triggered, this, callback);
     }
-
-private:
-
-    static void update_geometry_impl(StaticMesh const & mh, Qt3DCore::QGeometry * geom);
-
-    Qt3DCore::QGeometry * m_geometry = nullptr;
-    Qt3DRender::QGeometryRenderer * m_renderer = nullptr;
-    Qt3DRender::QMaterial * m_material = nullptr;
-
-}; /* end class RStaticMesh */
+}
 
 } /* end namespace modmesh */
