+++ cpp/modmesh/pilot/RAxisMark.cpp
@@ -26,8 +26,8 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <modmesh/view/common_detail.hpp> // Must be the first include.
-#include <modmesh/view/RAxisMark.hpp>
+#include <modmesh/pilot/common_detail.hpp> // Must be the first include.
+#include <modmesh/pilot/RAxisMark.hpp>
 
 #include <modmesh/modmesh.hpp>
