+++ cpp/modmesh/pilot/RCameraController.hpp
@@ -28,7 +28,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <modmesh/view/common_detail.hpp> // Must be the first include.
+#include <modmesh/pilot/common_detail.hpp> // Must be the first include.
 #include <QCamera>
 #include <QAbstractCameraController>
