+++ cpp/modmesh/pilot/RWorld.cpp
@@ -26,9 +26,9 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <modmesh/view/RWorld.hpp> // Must be the first include.
+#include <modmesh/pilot/RWorld.hpp> // Must be the first include.
 
-#include <modmesh/view/common_detail.hpp>
+#include <modmesh/pilot/common_detail.hpp>
 
 namespace modmesh
 {
