+++ resources/pilot/solvcon.icns
